move testimonials into a hook; update classbased component to functional
