+++ src/Hooks/useGetTestimonials.ts
@@ -1,0 +1,13 @@
+export default function useGetTestimonials() {
+  const testimonials = [
+    'You make my job easy.', // Eric
+    'I miss working with you.', // David
+    'I trust you to put your styling magic on it.', // Neil
+    'Excellent work as always, Shmoo.', // Hugo
+    "Don't forget to drink some water!", // David
+    'The CSS Wizard of our team!', // Johanna
+    'You absolutely can do it!', // Hugo
+  ]
+
+  return testimonials
+}
