+++ src/header/Random.tsx
@@ -1,15 +1,8 @@
 import { useState, useEffect } from 'react'
+import useGetTestimonials from '../Hooks/useGetTestimonials'
 
 export default function Random() {
-  const testimonials = [
-    'You make my job easy.', // Eric
-    'I miss working with you.', // David
-    'I trust you to put your styling magic on it.', // Neil
-    'Excellent work as always, Shmoo.', // Hugo
-    "Don't forget to drink some water!", // David
-    'The CSS Wizard of our team!', // Johanna
-    'You absolutely can do it!', // Hugo
-  ]
+  const testimonials = useGetTestimonials()
 
   const [testimonialIndex, setTestimonialIndex] = useState(0)
